@@ -10,26 +10,27 @@
   },
   "dependencies": {
     "axios": "^0.21.0",
-    "element-plus": "^v1.0.1-beta.17",
+    "element-plus": "^1.0.1-beta.26",
     "js-md5": "^0.7.3",
     "lodash": "^4.17.20",
     "mitt": "^2.1.0",
     "moment": "^2.29.1",
-    "vue": "^3.0.4",
-    "vue-axios": "^3.2.0",
-    "vue-router": "4.0.1",
+    "vue": "^3.0.5",
+    "vue-axios": "^3.2.2",
+    "vue-router": "4.0.3",
     "vuedraggable": "^4.0.1",
     "vuex": "4.0.0-rc.2"
   },
   "devDependencies": {
     "@types/js-md5": "^0.4.2",
     "@types/node": "^14.14.16",
-    "@vue/compiler-sfc": "^3.0.2",
-    "@vuedx/typecheck": "^0.2.4-0",
-    "@vuedx/typescript-plugin-vue": "^0.2.4-0",
+    "@vitejs/plugin-vue": "^1.0.4",
+    "@vue/compiler-sfc": "^3.0.5",
+    "@vuedx/typecheck": "^0.4.1",
+    "@vuedx/typescript-plugin-vue": "^0.4.1",
     "node-sass": "^5.0.0",
     "sass": "^1.29.0",
-    "typescript": "^4.0.3",
-    "vite": "^1.0.0-rc.13"
+    "typescript": "^4.1.3",
+    "vite": "^2.0.0-beta.29"
   }
 }
